perf(bundle): node-timeseries-map + synthetic generator off eager bundle
Two more transitive eager pulls hidden behind innocuous-looking imports.

1. RiskPropagationFlow (critical path) imported getNodeDataDescription
   from real-timeseries.ts. That helper just reads a constant map but
   real-timeseries.ts eagerly pulls NODE_TIMESERIES_MAP (1311 LOC of
   timeseries source definitions). Net ~1700 LOC of timeseries data
   rode the critical path for tiny 6px data-label badges.

   Fix: effect-loaded state pattern. The render site uses
   getNodeDataDescription?.(card.nodeId) so the badges are absent
   until the chunk lands (~50-100ms). Visual impact is negligible.

2. useTemporalGraph (used by RiskPropagationFlow, ModulePanel, TimeDial
   — everywhere with a timeline cursor) imported four pure reader
   functions from temporal-data.ts. The file also contained
   generateTemporalData (~145 LOC synthetic-data generator) which got
   pulled along even though it's only invoked once per session inside
   the dynamic-loaded initTemporalData.

   Fix: new temporal-state-helpers.ts (~130 LOC) houses the readers
   + the seven shared types. temporal-data.ts re-exports for backward
   compat. useTemporalGraph imports from the lightweight file directly.

vitest 1524/1524 pass. tsc clean.
